requirements.txt
Jinja2==2.7.3
Markdown==2.5.2
PyYAML==3.11
docopt==0.6.2
python-slugify==0.1.0
prettytable==0.7.2
six==1.8.0
testfixtures==4.1.2